FNN_Regression_Diabetes/requirements.txt
# requirements.txt
# 深度学习回归实验环境依赖

# 核心机器学习框架
torch>=2.0.0
scikit-learn>=1.0.0
numpy>=1.21.0

# 数据处理和可视化
matplotlib>=3.5.0
pyyaml>=6.0

# 可选但推荐
jupyter>=1.0.0
ipython>=8.0.0
tqdm>=4.60.0
